+++ lib/glimmer/swt/widget_proxy.rb
@@ -441,4 +441,4 @@ module Glimmer
       end
     end
   end
-end
+end