+++ lib/glimmer/ui/custom_shell.rb
@@ -8,7 +8,7 @@ module Glimmer
 
       def initialize(parent, *swt_constants, options, &content)
         super
-        raise Error, 'Invalid custom shell body root! Must be a shell or another custom shell.' unless body_root.is_a?(SWT::ShellProxy) || body_root.is_a?(CustomShell)
+        raise Error, 'Invalid custom shell body root! Must be a shell or another custom shell.' unless body_root.swt_widget.is_a?(org.eclipse.swt.widgets.Shell)
       end
 
       # Classes may override